# openapi.yaml
openapi: 3.0.3
info:
  title: AyudaBesh API
  description: |
    A service marketplace platform connecting customers with verified local providers.
    Supports user authentication, service booking, availability management, reviews,
    disputes, reports, notifications, and full admin moderation.
  version: 1.0.0
  contact:
    name: Alice (BSIT Student)
    email: alice@example.com
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: "User login"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security: []
  /api/auth/signup:
    post:
      tags: [Authentication]
      summary: "User registration"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: "Signup successful"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
      security: []
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: "User logout"
      responses:
        '200':
          description: "Logged out successfully"
      security:
        - JWTAuth: []
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: "Request password reset (sends verification code) - supports customer, provider, and admin roles"
      description: "Request a password reset for any user role. Admins can use this endpoint to reset their password."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: "Verification code sent"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
  /api/auth/admin/signup:
    post:
      tags: [Authentication]
      summary: "Create admin account (first admin public, subsequent require admin auth)"
      description: |
        Create a new admin account. 
        - If no admins exist: Public endpoint (no authentication required) - for initial setup
        - If admins exist: Requires admin authentication via JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminRequest'
      responses:
        '201':
          description: "Admin account created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: "Admin authentication required (when admins already exist)"
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: "Access denied - not an admin"
          $ref: '#/components/responses/Forbidden'
      security: []
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: "Reset password with verification code"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: "Password reset successful"
        '400':
          $ref: '#/components/responses/BadRequest'
      security: []

  # =============== USERS ===============
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Include a valid JWT token in the `Authorization: Bearer <token>` header."
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Access
      description: "Implicit — admin role is verified via JWT payload."
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [customer, provider, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, username, fullName, email, role]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, provider, admin]
      required: [username, password, role]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required: [user, token]
    SignupRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [customer, provider]
        services_offered:
          type: array
          items:
            type: string
        location:
          type: string
        description:
          type: string
        hourly_rate:
          type: number
        service_radius:
          type: number
        equipment:
          type: string
      required: [username, email, phone, password, fullName]
    ForgotPasswordRequest:
      type: object
      properties:
        identifier:
          type: string
          description: "email, phone, or username"
        role:
          type: string
          enum: [customer, provider, admin]
      required: [identifier]
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
        reset_token:
          type: string
        identifier_masked:
          type: string
        sent_via:
          type: array
          items:
            type: string
        verification_code:
          type: string
          description: "Only in dev mode (if email/SMS not configured)"
    ResetPasswordRequest:
      type: object
      properties:
        reset_token:
          type: string
        verification_code:
          type: string
        new_password:
          type: string
      required: [reset_token, verification_code, new_password]
    CreateAdminRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        fullName:
          type: string
      required: [username, email, phone, password, fullName]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
  responses:
    BadRequest:
      description: "Bad request (missing/invalid data)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: "Authentication required"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: "Access denied (e.g., wrong role)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - JWTAuth: []